# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sakura Beauty E-Commerce API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MessageResponse {
  message: string;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  /** @nullable */
  icon: string | null;
  displayOrder: number;
  createdAt: string;
  updatedAt: string;
}

export interface CreateCategoryBody {
  name: string;
  slug?: string;
  /** @nullable */
  icon?: string | null;
  displayOrder?: number;
}

export type ProductMainIngredientsItem = {
  name: string;
  icon: string;
};

export interface Product {
  id: number;
  name: string;
  slug: string;
  price: number;
  /** @nullable */
  discountPrice: number | null;
  category: string;
  stock: number;
  description: string;
  /** @nullable */
  ingredients: string | null;
  keyBenefits: string[];
  mainIngredients: ProductMainIngredientsItem[];
  bestFor: string[];
  /** @nullable */
  texture: string | null;
  images: string[];
  averageRating: number;
  reviewCount: number;
  isFeatured: boolean;
  /** @nullable */
  homepageSection: string | null;
  createdAt: string;
}

export interface ProductListResponse {
  products: Product[];
  total: number;
  page: number;
  totalPages: number;
}

export interface MainIngredient {
  name: string;
  icon: string;
}

export interface CreateProductBody {
  name: string;
  price: number;
  /** @nullable */
  discountPrice?: number | null;
  category: string;
  stock: number;
  description: string;
  /** @nullable */
  homepageSection?: string | null;
  /** @nullable */
  ingredients?: string | null;
  keyBenefits?: string[];
  mainIngredients?: MainIngredient[];
  bestFor?: string[];
  /** @nullable */
  texture?: string | null;
  images: string[];
  isFeatured?: boolean;
}

export interface UpdateProductBody {
  name?: string;
  price?: number;
  /** @nullable */
  discountPrice?: number | null;
  category?: string;
  stock?: number;
  description?: string;
  /** @nullable */
  ingredients?: string | null;
  keyBenefits?: string[];
  mainIngredients?: MainIngredient[];
  bestFor?: string[];
  /** @nullable */
  texture?: string | null;
  images?: string[];
  isFeatured?: boolean;
}

export interface CartItem {
  id: number;
  productId: number;
  quantity: number;
  product: Product;
}

export interface Cart {
  items: CartItem[];
  subtotal: number;
  discount: number;
  total: number;
}

export interface AddToCartBody {
  productId: number;
  quantity: number;
}

export interface UpdateCartItemBody {
  quantity: number;
}

export interface OrderItem {
  id: number;
  productId: number;
  productName: string;
  productImage: string;
  quantity: number;
  price: number;
}

export interface AddressBody {
  fullName: string;
  phone: string;
  street: string;
  city: string;
  district: string;
  /** @nullable */
  postalCode?: string | null;
  isDefault?: boolean;
}

export interface Order {
  id: number;
  trackingId: string;
  userId: string;
  items: OrderItem[];
  totalAmount: number;
  paymentMethod: string;
  paymentStatus: string;
  orderStatus: string;
  /** @nullable */
  transactionId: string | null;
  shippingAddress: AddressBody;
  /** @nullable */
  couponCode: string | null;
  discountAmount: number;
  /** @nullable */
  cancellationReason?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface TrackingEvent {
  status: string;
  label: string;
  /** @nullable */
  timestamp: string | null;
  completed: boolean;
}

export interface OrderTracking {
  trackingId: string;
  orderStatus: string;
  paymentStatus: string;
  paymentMethod: string;
  createdAt: string;
  updatedAt: string;
  timeline: TrackingEvent[];
}

export interface CreateOrderBody {
  paymentMethod: string;
  /** @nullable */
  transactionId?: string | null;
  shippingAddress: AddressBody;
  /** @nullable */
  couponCode?: string | null;
}

export interface UpdateOrderStatusBody {
  orderStatus: string;
  /** @nullable */
  cancellationReason?: string | null;
}

export interface UpdatePaymentStatusBody {
  paymentStatus: string;
}

export type ReviewEligibilityReason =
  | (typeof ReviewEligibilityReason)[keyof typeof ReviewEligibilityReason]
  | null;

export const ReviewEligibilityReason = {
  already_reviewed: "already_reviewed",
  not_purchased: "not_purchased",
} as const;

export interface ReviewEligibility {
  canReview: boolean;
  reason?: ReviewEligibilityReason;
}

export interface Review {
  id: number;
  productId: number;
  userId: string;
  userName: string;
  rating: number;
  comment: string;
  createdAt: string;
}

export interface CreateReviewBody {
  rating: number;
  comment: string;
}

export interface AdminReview {
  id: number;
  productId: number;
  userId: string;
  userName: string;
  rating: number;
  comment: string;
  createdAt: string;
  productName: string;
  productImage?: string | null;
}

export interface WishlistItem {
  id: number;
  productId: number;
  product: Product;
  addedAt: string;
}

export interface Coupon {
  id: number;
  code: string;
  discountType: string;
  discountValue: number;
  /** @nullable */
  minOrderAmount: number | null;
  /** @nullable */
  expiryDate: string | null;
  isActive: boolean;
}

export interface ValidateCouponBody {
  code: string;
  orderAmount: number;
}

export interface CreateCouponBody {
  code: string;
  discountType: string;
  discountValue: number;
  /** @nullable */
  minOrderAmount?: number | null;
  /** @nullable */
  expiryDate?: string | null;
}

export interface UserProfile {
  id: number;
  clerkId: string;
  email: string;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  phone: string | null;
  role: string;
  isBlocked: boolean;
  createdAt: string;
}

export interface UpdateProfileBody {
  firstName?: string;
  lastName?: string;
  phone?: string;
}

export interface Address {
  id: number;
  userId: string;
  fullName: string;
  phone: string;
  street: string;
  city: string;
  district: string;
  /** @nullable */
  postalCode: string | null;
  isDefault: boolean;
}

export interface CategorySales {
  category: string;
  total: number;
  count: number;
}

export interface MonthlySales {
  month: string;
  total: number;
  orders: number;
}

export interface DashboardStats {
  totalSales: number;
  totalOrders: number;
  totalUsers: number;
  pendingOrders: number;
  recentOrders: Order[];
  salesByCategory: CategorySales[];
  monthlySales: MonthlySales[];
}

export interface ToggleBlockBody {
  isBlocked: boolean;
}

export type ListProductsParams = {
  category?: string;
  search?: string;
  minPrice?: number;
  maxPrice?: number;
  minRating?: number;
  sortBy?: string;
  page?: number;
  limit?: number;
};

export type GetHomepageProducts200 = {
  top: Product[];
  bottom: Product[];
};

export type ListAllOrdersParams = {
  status?: string;
  page?: number;
};