# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Sakura Beauty E-Commerce API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all categories
      responses:
        "200":
          description: Categories list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create a category (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryBody"
      responses:
        "201":
          description: Created category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}:
    put:
      operationId: updateCategory
      tags: [categories]
      summary: Update a category (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryBody"
      responses:
        "200":
          description: Updated category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete a category (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /cart/items:
    post:
      operationId: addToCart
      tags: [cart]
      summary: Add item to cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddToCartBody"
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
  /cart/items/{productId}:
    put:
      operationId: updateCartItem
      tags: [cart]
      summary: Update cart item quantity
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCartItemBody"
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
    delete:
      operationId: removeFromCart
      tags: [cart]
      summary: Remove item from cart
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
...
  /admin/orders:
    get:
      operationId: listAllOrders
      tags: [admin]
      summary: List all orders (admin only)
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: All orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
  /admin/orders/{id}/status:
    put:
      operationId: updateOrderStatus
      tags: [admin]
      summary: Update order status (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderStatusBody"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /admin/orders/{id}/payment:
    put:
      operationId: updateOrderPayment
      tags: [admin]
      summary: Verify/update order payment status (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePaymentStatusBody"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
...
  /admin/users/{id}/block:
    put:
      operationId: toggleUserBlock
      tags: [admin]
      summary: Block/unblock user (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ToggleBlockBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        icon:
          type: ["string", "null"]
        displayOrder:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - slug
        - icon
        - displayOrder
        - createdAt
        - updatedAt
    CreateCategoryBody:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        icon:
          type: ["string", "null"]
        displayOrder:
          type: integer
      required:
        - name
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        price:
          type: number
        discountPrice:
          type: ["number", "null"]
        category:
          type: string
        stock:
          type: integer
        description:
          type: string
        ingredients:
          type: ["string", "null"]
        keyBenefits:
          type: array
          items:
            type: string
        mainIngredients:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              icon:
                type: string
            required: [name, icon]
        bestFor:
          type: array
          items:
            type: string
        texture:
          type: ["string", "null"]
        images:
          type: array
          items:
            type: string
        averageRating:
          type: number
        reviewCount:
          type: integer
        isFeatured:
          type: boolean
        homepageSection:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - name
        - slug
        - price
        - discountPrice
        - category
        - stock
        - description
        - ingredients
        - keyBenefits
        - mainIngredients
        - bestFor
        - texture
        - images
        - averageRating
        - reviewCount
        - isFeatured
        - homepageSection
        - createdAt
...
    CartItem:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        quantity:
          type: integer
        product:
          $ref: "#/components/schemas/Product"
      required:
        - id
        - productId
        - quantity
        - product
    Cart:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
        subtotal:
          type: number
        discount:
          type: number
        total:
          type: number
      required:
        - items
        - subtotal
        - discount
        - total
    AddToCartBody:
      type: object
      properties:
        productId:
          type: integer
        quantity:
          type: integer
      required:
        - productId
        - quantity
    UpdateCartItemBody:
      type: object
      properties:
        quantity:
          type: integer
      required:
        - quantity
    OrderItem:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        productImage:
          type: string
        quantity:
          type: integer
        price:
          type: number
      required:
        - id
        - productId
        - productName
        - productImage
        - quantity
        - price
    Order:
      type: object
      properties:
        id:
          type: integer
        trackingId:
          type: string
        userId:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        totalAmount:
          type: number
        paymentMethod:
          type: string
        paymentStatus:
          type: string
        orderStatus:
          type: string
        transactionId:
          type: ["string", "null"]
        shippingAddress:
          $ref: "#/components/schemas/AddressBody"
        couponCode:
          type: ["string", "null"]
        discountAmount:
          type: number
        cancellationReason:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - trackingId
        - userId
        - items
        - totalAmount
        - paymentMethod
        - paymentStatus
        - orderStatus
        - transactionId
        - shippingAddress
        - couponCode
        - discountAmount
        - createdAt
        - updatedAt
...
    UpdateOrderStatusBody:
      type: object
      properties:
        orderStatus:
          type: string
        cancellationReason:
          type: ["string", "null"]
      required:
        - orderStatus
    UpdatePaymentStatusBody:
      type: object
      properties:
        paymentStatus:
          type: string
      required:
        - paymentStatus
...
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        clerkId:
          type: string
        email:
          type: string
        firstName:
          type: ["string", "null"]
        lastName:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        role:
          type: string
        isBlocked:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - clerkId
        - email
        - firstName
        - lastName
        - phone
        - role
        - isBlocked
        - createdAt
...
    AddressBody:
      type: object
      properties:
        fullName:
          type: string
        phone:
          type: string
        street:
          type: string
        city:
          type: string
        district:
          type: string
        postalCode:
          type: ["string", "null"]
        isDefault:
          type: boolean
      required:
        - fullName
        - phone
        - street
        - city
        - district
...
    ToggleBlockBody:
      type: object
      properties:
        isBlocked:
          type: boolean
      required:
        - isBlocked